LMDB

Exported

Base.Iterators.drop — Method

Empty or delete+close a database.

If parameter delete is false DB will be emptied, otherwise DB will be deleted from the environment and DB handle will be closed

Base.close — Method

Close a cursor

Base.close — Method

Close a database handle

Base.close — Method

Close the environment and release the memory map

Base.count — Method

Return count of duplicates for current key

Base.delete! — Function

Delete items from a database

Base.delete! — Method

Delete current key/data pair to which the cursor refers

Base.get — Method

Retrieve by cursor.

This function retrieves key/data pairs from the database.

Base.get — Method

Get items from a database

Base.getindex — Method

Get environment flags and parameters

getindex accepts folowing parameters:

env db environment object

option symbol which indicates parameter. Currently supported parameters:

Flags

Readers

KeySize

Note: Consult LMDB documentation for particual values of environment parameters and flags.

Base.isopen — Method

Check if cursor is open

Base.isopen — Method

Check if database is open

Base.isopen — Method

Check if environment is open

Base.isopen — Method

Check if transaction is open.

Base.keys — Method

Return iterator over keys of uniform, specified type

Base.open — Function

Open a database in the environment

Base.open — Function

Wrapper of DBI open for do construct

Base.open — Method

Open an environment handle

open function accepts folowing parameters:

env db environment object

path directory in which the database files reside

flags defines special options for the environment

mode UNIX permissions to set on created files

Note: A database directory must exist and be writable.

Base.open — Method

Wrapper of Cursor open for do construct

Base.open — Method

Create a cursor

Base.put! — Method

Store by cursor.

This function stores key/data pairs into the database. The cursor is positioned at the new item, or on failure usually near it.

Base.put! — Method

Store items into a database

Base.reset — Method

Reset a read-only transaction

Abort the transaction like abort, but keep the transaction handle.

Base.setindex! — Method

Set environment flags and parameters

setindex! accepts folowing parameters:

env db environment object

option symbol which indicates parameter. Currently supported parameters:

Flags

Readers

MapSize

DBs

value parameter value

Note: Consult LMDB documentation for particual values of environment parameters and flags.

LMDB.abort — Method

Abandon all the operations of the transaction instead of saving them

The transaction and its cursors must not be used after, because its handle is freed.

LMDB.commit — Method

Commit all the operations of a transaction into the database

The transaction and its cursors must not be used after, because its handle is freed.

LMDB.create — Method

Wrapper of create for do construct

LMDB.create — Method

Create an LMDB environment handle

LMDB.database — Method

Return the cursor's database

LMDB.environment — Method

Wrapper of open for do construct

LMDB.info — Method

Return information about the LMDB environment.

LMDB.isflagset — Method

Check if binary flag is set in provided value

LMDB.path — Method

Return the path that was used in open

LMDB.renew — Method

Renew a cursor

LMDB.renew — Method

Renew a read-only transaction

This acquires a new reader lock for a transaction handle that had been released by reset. It must be called before a reset transaction may be used again.

LMDB.set! — Method

Set environment flags

LMDB.start — Method

Create a transaction for use with the environment

start function creates a new transaction and returns Transaction object. It allows to set transaction flags with flags option.

LMDB.sync — Function

Flush the data buffers to disk

LMDB.transaction — Method

Return the cursor's transaction

LMDB.unset! — Method

Unset environment flags

LMDB.Cursor — Type

A handle to a cursor structure for navigating through a database.

LMDB.DBI — Type

A handle for an individual database in the DB environment.

LMDB.Environment — Type

A DB environment supports multiple databases, all residing in the same shared-memory map.

LMDB.LMDBError — Type

LMDB exception type

LMDB.Transaction — Type

A database transaction. Every operation requires a transaction handle. All database operations require a transaction handle. Transactions may be read-only or read-write.

Not Exported

Base.iterate — Method

Iterate over database

LMDB.errormsg — Method

Return a string describing a given error code

Function returns description of the error as a string. It accepts following arguments:

err::Int32: An error code.

LMDB.flags — Method

Retrieve the DB flags for a database handle

LMDB.version — Method

Return the LMDB library version and version information

Function returns tuple (VersionNumber, String) that contains a library version and a library version string.

LMDB.LMDBIterator — Type

Type to implement the Iterator interface